requirements.txt
python-binance==1.0.19
pandas==2.0.3
numpy==1.24.3
colorama==0.4.6
